Run  on entire dataset rather than filtering for NFE samples
