+++ scripts/hail_batch/hgdp1kg_tobwgs_nfe_pc_relate/main.py
@@ -8,15 +8,15 @@ service_backend = hb.ServiceBackend(
     billing_project=os.getenv('HAIL_BILLING_PROJECT'), bucket=os.getenv('HAIL_BUCKET')
 )
 
-batch = hb.Batch(name='nfe-pc-relate', backend=service_backend)
+batch = hb.Batch(name='pc-relate', backend=service_backend)
 
 dataproc.hail_dataproc_job(
     batch,
-    f'hgdp_1kg_tob_wgs_nfe_pc_relate.py',
+    f'hgdp_1kg_tob_wgs_pc_relate.py',
     max_age='4h',
     num_secondary_workers=20,
     init=['gs://cpg-reference/hail_dataproc/install_common.sh'],
-    job_name=f'nfe-pc-relate',
+    job_name=f'pc-relate',
     worker_boot_disk_size=200,
 )
